+++ src/index.ts
@@ -1,4 +1,5 @@
-export { Spec, Operation, Claim, Constant } from './program-spec.ts';
+export { Spec, Claim, Constant } from './program-spec.ts';
+export { Operation } from './operation.ts';
 export { Credential } from './credential-index.ts';
 export { Presentation, PresentationRequest } from './presentation.ts';
 export { assert } from './util.ts';
